Baseclass for UserLists- and UserVariableBricks
*update of tests which use Reflection.
